+++ all/src/main/java/module-info.java
@@ -1,10 +1,12 @@
 module no.beint.glimt.all {
     requires transitive no.beint.glimt.avif;
     requires transitive no.beint.glimt.jpeg;
+    requires transitive no.beint.glimt.jpegli;
     requires transitive no.beint.glimt.png;
     requires transitive no.beint.glimt.webp;
     requires transitive no.beint.glimt.heic;
     requires transitive no.beint.glimt.jxl;
     requires transitive no.beint.glimt.extra;
     requires transitive no.beint.glimt.imageio;
+    requires transitive no.beint.glimt.resize;
 }